Dicke.DickeHusimiProjections

Dicke.DickeHusimiProjections.matrix_QP∫∫dqdpδϵ — Method

function matrix_QP∫∫dqdpδϵ(system::ClassicalDickeSystem;
    f::Function,
    ϵ::Real,
    res::Real=0.1,
    symmetricQP::Bool=false,
    symmetricP::Bool=symmetricQP,
    paralelize::Bool=(Distributed.myid()==1),
    showprogress::Bool = true,
    pbatch_size::Integer=Int(min(ceil((4/res)^2/Distributed.nprocs()/10),50)))

Returns a tuple (Qs, Ps, mat) where Qs and Ps are real vectors spanning all possible values of $P$ and $Q$ with some step res, and mat is a matrix whose entries are given by mat[i,j] =∫∫dqdpδϵ(system; ϵ=ϵ, Q=Qs[i], P=P[j], p_res=res,kargs...)

Arguments

system should be an instance of ClassicalDicke.ClassicalDickeSystem.

ϵ is a real number (see arguments for ∫∫dqdpδϵ).

f should be function with signature f([Q,q,P,p]) that returns values that may be added together (Numbers, Arrays etc...).

res determines the separation between the elements of Qs and Ps. It also determines p_res in the calls to ∫∫dqdpδϵ.

symmetricQP indicates that f([Q,q,P,p]) = f([-Q,q,P,p]) = f([Q,q,-P,p]) for all Q and P. If true only the non-positive entries of Q and P will be computed, and the other entries will be mirrored. Default is false.

symmetricP indicates that f([Q,q,-P,p]) = f([Q,q,P,p]) for all P. If true, only the side with non-positive P coordinate will be comptued, and the other entries will be mirrored. The default is to be the same that symmetricQP.

paralelize indicates whether to use all available workers. Defaults to true if this function is called from worker 1, and false else.

showprogress turns the progress bar on/off. Default is true.

pbatch_size is the maximum number of batches to run in a single run in a single worker. The default value is automatically optimized depending on res and the number of workers.

kargs... are redirected to ∫∫dqdpδϵ.

Dicke.DickeHusimiProjections.∫∫dqdpδϵ — Function

function ∫∫dqdpδϵ(system::ClassicalDickeSystem;
    ϵ::Real,
    Q::Real,
    P::Real,
    f::Function,
    p_res::Real=0.01,
    nonvalue=nothing,
    onlyqroot::Union{typeof(-),typeof(+),Nothing}=nothing)

Computes the integral

\[ \int \int dq dp f(\mathbf{x}) \delta(h_\text{cl}(\mathbf{x}) - \epsilon)\]

using Eq. (8) of Ref. [11] and a Chebyshev–Gauss quadrature method.

Note: to type ∫ in Julia, type \int + Tab.

Arguments

system should be an instance of ClassicalDicke.ClassicalDickeSystem.

ϵ is a real number corresponding to $\epsilon$ above.

Q and P are the $Q$ and $P$ coordinates of $\mathbf{x}$ above.

f should be function with signature f([Q,q,P,p]) that returns values that may be added together (Numbers, Arrays etc...).

p_res is the resolution of the integral. The amount of nodes for the quadrature will be equal to 2*floor(p₊/p_res) + 1, where p₊ is defined below Eq. (9) of Ref. [11]. Default is 0.01.

nonvalue is the value to return if $Q,P$ are outside of the energy shell at $\epsilon$. Default is NaN.

onlyqroot may be +, -, or nothing (default). If it is + (-), the integral is only computed over the positive (negative) roots in q, that is, only the $q_+$ ($q_-$) term is taken in Eq. (8) of Ref. [11]. If it is nothing, both terms are added.

Dicke.DickeHusimiProjections.∫∫dqdpδϵ_Husimi_Renyi_powers — Method

function ∫∫dqdpδϵ_Husimi_Renyi_powers(
    systemQ::DickeBCE.QuantumDickeSystem;
    states::Union{Array{<:Number,1},Array{<:Number,2}},
    res::Real,
    Htol::Real=1e-4,
    nonvalue=NaN,
    averageallstates::Bool=false,
    averagingfunction::Function=mean,
    α::Union{AbstractArray{<:Real,1},Real}=2,
    kargs...)

Calls matrix_QP∫∫dqdpδϵ PENDING