DickeModel.PhaseSpaces

This module provides formulas for the cannonical transformation $(Q,P)\mapsto (\theta,\phi)$ we use throughout this package.

Note

There is an unfortunate problem between the characters $\phi$ (\phi) and $\varphi$ (\varphi): some fonts render one as the other. To avoid errors, the functions below treat the two characters indistinctly. In the references, the glyph $\phi$ is used.

DickeModel.PhaseSpaces.P_of_θφ — Method

See P_of_θϕ.

DickeModel.PhaseSpaces.P_of_θϕ — Method

function P_of_θϕ(θ,ϕ)

Returns

\[ P=- \sqrt{2 (1 - \cos(\theta))}\sin(\phi).\]

Also P_of_θφ(θ,φ).

DickeModel.PhaseSpaces.Q_of_θφ — Method

See Q_of_θϕ.

DickeModel.PhaseSpaces.Q_of_θϕ — Method

function Q_of_θϕ(θ,ϕ)

Returns

\[ Q=\sqrt{2 (1 - \cos(\theta))}\cos(\phi).\]

Also Q_of_θφ(θ,φ).

DickeModel.PhaseSpaces.arc_between_QP — Method

function arc_between_QP(Q1,P1,Q2,P2)

Returns

    arc_between_θϕ(θ_of_QP(Q1,P1),ϕ_of_QP(Q1,P1),θ_of_QP(Q2,P2),ϕ_of_QP(Q2,P2)).

DickeModel.PhaseSpaces.arc_between_θφ — Method

See arc_between_θϕ.

DickeModel.PhaseSpaces.arc_between_θϕ — Method

function arc_between_θϕ(θ1,φ1,θ2,φ2)

Returns

\[ \Theta = \arccos(\cos(\theta_1)\cos(\theta_2)+ \sin(\theta_1)\sin(\theta_2)\cos(\phi_1 - \phi_2)).\]

Also arc_between_θφ(θ1,φ1,θ2,φ2).

DickeModel.PhaseSpaces.jx — Method

function jx(Q,P)

Returns

\[ j_x = \sin(\theta(Q,P))\cos(\phi(Q,P)).\]

DickeModel.PhaseSpaces.jy — Method

function jy(Q,P)

Returns

\[ j_y = \sin(\theta(Q,P))\sin(\phi(Q,P)).\]

DickeModel.PhaseSpaces.jz — Method

function jz(Q,P)

Returns

\[ j_z = - \cos(\theta(Q,P)).\]

where $\text{arctan2}$ is the 2-argument arctangent.

DickeModel.PhaseSpaces.θ_of_QP — Method

function θ_of_QP(Q,P)

Returns

\[ \theta =- \arccos\left(1 - \frac{Q^2 + P^2}{2}\right).\]

DickeModel.PhaseSpaces.φ_of_QP — Method

See ϕ_of_QP.

DickeModel.PhaseSpaces.ϕ_of_QP — Method

function ϕ_of_QP(Q,P)

Returns

\[ \phi = \text{arctan2}(-P,Q)\in [0,2 \pi ],\]

where $\text{arctan2}$ is the 2-argument arctangent.

Also φ_of_QP(Q,P).