Dicke.ClassicalDicke

Dicke.ClassicalDicke.density_of_states — Method

function density_of_states(system::ClassicalSystems.ClassicalSystem;j,ε)

Returns the semiclassical density of states (DoS) $ν(ϵ)$, in units of $1/ϵ$. This is computed with an expression similar to Eq. (19) of 10.1103/PhysRevA.89.032101, where we add an additional factor of $j$ to have units of $1/ϵ$ instead of $1/E$, and the integral is performed with a change of variable.

Arguments:

system is an instance of ClassicalSystems.ClassicalSystem.

j is the value of $j$

ε is the scaled energy $ϵ=E/j$

Dicke.ClassicalDicke.energy_shell_volume — Method

function energy_shell_volume(system::ClassicalSystems.ClassicalSystem;ε)

Returns the volume of the classical energy shell in the phase space, that is, $\mathcal{V}(\mathcal{M}_ϵ)$ in Eq. (27) of 10.1103/PhysRevE.103.052214.

Arguments:

system is an instance of ClassicalSystems.ClassicalSystem.

ε is the scaled energy $ϵ=E/j$

Dicke.ClassicalDicke.hamiltonian — Method

function hamiltonian(system::ClassicalSystems.ClassicalSystem)

Returns a classical Hamiltonian function h(x) where x=[Q,q,P,p], which is given by Eq. (5) of 10.1088/1367-2630/abd2e6.

Arguments:

system is an instance of ClassicalSystems.ClassicalSystem.

Dicke.ClassicalDicke.maximum_P_for_ε — Method

maximumPfor_ε(sistema::ClassicalSystems.ClassicalSystem,ε) Computes the maximum value of the parameter P accessible to the system at energy ε.

Dicke.ClassicalDicke.maximum_Q_for_ε — Method

maximumQfor_ε(sistema::ClassicalSystems.ClassicalSystem,ε) Computes the maximum value of the parameter Q accessible to the system at energy ε.