Dicke.ClassicalDicke

Dicke.ClassicalDicke.ClassicalDickeSystem — Type

struct ClassicalDickeSystem <: ClassicalSystems.ClassicalSystem

Instance of ClassicalDicke.ClassicalDickeSystem which represents the classical Dicke model with the given parameters $ω_0$, $ω$, and $γ$. See Eq. (5) of Ref. [10]. To generate this struct, use the constructor

ClassicalDickeSystem(;ω₀::Real,ω::Real,γ::Real)

For example, system = ClassicalDickeSystem(;ω₀=1, ω=1, γ=1).

This struct may be passed to all functions in this module that require an instance of ClassicalDicke.ClassicalDickeSystem, as well as functions in other modules that require the abstract ClassicalSystems.ClassicalSystem, such as ClassicalSystems.integrate.

Dicke.ClassicalDicke.Point — Method

function Point(system::ClassicalSystems.ClassicalSystem;Q,P,p,ϵ,signo::Union{typeof(-),typeof(+)}=+)

Returns a list [Q,q,P,p], where q is calculated with ClassicalDicke.q_of_ϵ. See that function for details on the arguments. If there are no solutions for $q$, an error is raised.

Dicke.ClassicalDicke.Pointθφ — Method

function Pointθφ(;θ,φ,q,p)

Returns a list [Q,q,P,p], where Q and P are calculated from θ and φ using PhaseSpaces.Q_of_θφ and PhaseSpaces.P_of_θφ.

Dicke.ClassicalDicke.classicalPathRandomSampler — Method

function classicalPathRandomSampler(system::ClassicalDickeSystem;ϵ,dt=3)

This function returns a function sample(), which produces random points within the classical energy shell at energy ϵ. This only works at energy shells where the classical dynamics are ergodic. The function sample() returns points from a fixed chaotic trajectory picked at random separated by a fixed time interval dt.

Arguments:

system should be generated with ClassicalDicke.ClassicalDickeSystem.

ϵ is the scaled energy $ϵ=E/j$ of the energy shell from where to sample.

dt is the fixed time interval that separates the points that are returned by sample().

Dicke.ClassicalDicke.density_of_states — Method

function density_of_states(system::ClassicalDickeSystem;j,ϵ)

Returns the semiclassical density of states (DoS) $ν(ϵ)$, in units of $1/ϵ$. This is computed with an expression similar to Eq. (19) of Ref. [2], where we add an additional factor of $j$ to have units of $1/ϵ$ instead of $1/E$, and the integral is performed with a change of variable.

Arguments:

system should be generated with ClassicalDicke.ClassicalDickeSystem.

j is the value of $j$

ϵ is the scaled energy $ϵ=E/j$

See Plotting the density of states for an example.

Dicke.ClassicalDicke.discriminant_of_q_solution — Method

function discriminant_of_q_solution(system::ClassicalDickeSystem; Q,P,p,ϵ)

Returns the discriminant of the second degree equation in $q$ given by

\[ h_\text{cl}(Q,q,P,p)=\epsilon,\]

where $h_\text{cl}$ is given by Eq. (5) of Ref. [10].

Arguments:

system should be generated with ClassicalDicke.ClassicalDickeSystem.

Q, P, p, and ϵ are the values of $Q$, $P$, $p$, and $\epsilon$, respectively.

Dicke.ClassicalDicke.energy_shell_volume — Method

function energy_shell_volume(system::ClassicalDickeSystem;ϵ)

Returns the volume of the classical energy shell in the phase space, that is, $\mathcal{V}(\mathcal{M}_ϵ)$ in Eq. (27) of Ref. [9].

Arguments:

system should be generated with ClassicalDicke.ClassicalDickeSystem.

ϵ is the scaled energy $ϵ=E/j$

Dicke.ClassicalDicke.energy_width_of_coherent_state — Method

function energy_width_of_coherent_state(system::ClassicalDickeSystem,x,j::Real)

Returns the energy width $\sigma$ of the coherent state $\left | \mathbf{x}\right \rangle$, in units of $\epsilon$. This quantity is given by $\sigma_D/j$ with $\sigma_D$ as in App. A of Ref. [5].

Arguments:

system should be generated with ClassicalDicke.ClassicalDickeSystem.

x is the coordinate $\mathbf{x}$ of the coherent state in the format [Q,q,P,p].

j is the value of $j$.

Dicke.ClassicalDicke.hamiltonian — Method

function hamiltonian(system::ClassicalDickeSystem)

Returns a classical Hamiltonian function h(x) where x=[Q,q,P,p], which is given by Eq. (5) of Ref. [10].

Arguments:

system should be generated with ClassicalDicke.ClassicalDickeSystem.

Dicke.ClassicalDicke.maximum_P_for_ϵ — Method

function maximum_P_for_ϵ(system::ClassicalDickeSystem,ϵ)

Computes the maximum value of the parameter $P$ accessible to the system at energy $\epsilon$.

Arguments:

system should be generated with ClassicalDicke.ClassicalDickeSystem.

ϵ is the scaled energy $ϵ=E/j$.

Dicke.ClassicalDicke.maximum_Q_for_ϵ — Method

function maximum_Q_for_ϵ(system::ClassicalDickeSystem,ϵ)

See ClassicalDicke.maximum_P_for_ϵ.

Dicke.ClassicalDicke.minimum_energy — Method

function minimum_energy(system::ClassicalDickeSystem)

Returns the energy of the ground-state coordinate given by ClassicalDicke.minimum_energy_point.

Arguments:

system should be generated with ClassicalDicke.ClassicalDickeSystem.

Note: This function currently only works for the supperadiant phase.

Dicke.ClassicalDicke.minimum_energy_point — Function

function minimum_energy_point(system::ClassicalDickeSystem,Qsign::Union{typeof(-),typeof(+)}=+)

Returns the ground-state coordinate, that is, $\mathbf{x}_\text{GS}$ below Eq. (7) of Ref. [10].

Arguments:

system should be generated with ClassicalDicke.ClassicalDickeSystem.

Qsign toggles the sign of the $Q$ coordinate, that is, + for $\mathbf{x}_\text{GS}$ and - for $\widetilde{\mathbf{x}}_\text{GS}$.

Note: This function currently only works for the superradiant phase.

Dicke.ClassicalDicke.minimum_ϵ_for — Method

function minimum_ϵ_for(system::ClassicalDickeSystem;Q=:nothing,q=:nothing,P=:nothing,p=:nothing)

Returns the minimum energy $\epsilon$ when constraining the system to three fixed values of the coordinates $Q$, $q$, $P$, $p$.

Arguments:

system should be generated with ClassicalDicke.ClassicalDickeSystem.

You may pass either $(Q,q,P)$ or $(q,P,p)$. The other combinanations are not implemented.

This function can be especially useful to draw contours of the available phase space (see Drawing contours of the available phase space)

Dicke.ClassicalDicke.normal_frequency — Function

function normal_frequency(system::ClassicalDickeSystem,signo::Union{typeof(-),typeof(+)}=+)

Returns the ground-state normal frequency, that is, $\Omega_{\epsilon_{\text{GS}}}^{A,B}$ at the bottom of page 3 of Ref. [10].

Arguments:

system should be generated with ClassicalDicke.ClassicalDickeSystem.

signo is - for $\Omega^A$ and + for $\Omega^B$.

Note: This function currently only works for the supperadiant phase.

Dicke.ClassicalDicke.phase_space_dist_squared — Method

function phase_space_dist_squared(x,y)

Returns the phase-space distance $d_{\mathcal{M}}(\mathbf{x},\mathbf{y})$ (See App. C of Ref [10]), where x and y are in the form [Q,q,P,p].

Dicke.ClassicalDicke.q_of_ϵ — Method

function q_of_ϵ(system::ClassicalDickeSystem;Q,P,p,ϵ,signo::Union{typeof(-),typeof(+)}=+,returnNaNonError=true)

Returns the solutions $q_\pm$ of the second degree equation in $q$ given by

\[ h_\text{cl}(Q,q,P,p)=\epsilon,\]

where $h_\text{cl}$ is given by Eq. (5) of Ref. [10].

Arguments:

system should be generated with ClassicalDicke.ClassicalDickeSystem.

Q, P, p, and ϵ are values of $Q$, $P$, $p$, and $\epsilon$, respectively.

signo is + for $q_+$ and - for $q_-$

If returnNaNonError is true, then NaN is returned if there are no solutions. If it is false, and error is raised.

Dicke.ClassicalDicke.q_sign — Function

function q_sign(system::ClassicalDickeSystem,x,ϵ=hamiltonian(system)(x))

Returns the sign of the root of the second degree equation in $q$ given by

\[ h_\text{cl}(Q,q,P,p)=\epsilon.\]

That is, this function returns + if q=x[2] ≈ q_of_ϵ(system;Q=x[1],P=x[3],p=x[4],ϵ=ϵ,signo=+) and returns - if q=x[2] ≈ q_of_ϵ(system;Q=x[1],P=x[3],p=x[4],ϵ=ϵ,signo=-).

Arguments:

system should be generated with ClassicalDicke.ClassicalDickeSystem.

x is a vector in the form [Q,q,P,p].

ϵ should be the energy of x. If this is not passed, it is computed using hamiltonian(system).