Dicke.ClassicalDicke

Dicke.ClassicalDicke.ClassicalSystem — Method

function ClassicalSystem(;ω₀,ω,γ)

Generates an instance of ClassicalSystems.ClassicalSystem which represents the classical Dicke model with the given parameters $ω_0$, $ω$, and $γ$. See Eq. (5) of Ref. Saúl Pilatowsky-Cameo, David Villaseñor, Miguel A. Bastarrachea-Magnani, Sergio Lerma-Hernández, Lea F. Santos, Jorge G. Hirsch (2021).

The returned value may be passed to all functions in this module that require an instance of ClassicalSystems.ClassicalSystem, as well as functions in other modules, such as ClassicalSystems.integrate.

Dicke.ClassicalDicke.Point — Method

function Point(system::ClassicalSystems.ClassicalSystem;Q,P,p,ε,signo::Union{typeof(-),typeof(+)}=+)

Returns a list [Q,q,P,p], where q is calculated with ClassicalDicke.q_of_ε. See that function for details on the arguments. If there are no solutions for $q$, an error is raised.

Dicke.ClassicalDicke.Pointθϕ — Method

function Pointθϕ(;θ,ϕ,q,p)

Returns a list [Q,q,P,p], where Q and P are calculated from θ and ϕ using PhaseSpaces.Q_of_θϕ and PhaseSpaces.P_of_θϕ.

Dicke.ClassicalDicke.density_of_states — Method

function density_of_states(system::ClassicalSystems.ClassicalSystem;j,ε)

Returns the semiclassical density of states (DoS) $ν(ϵ)$, in units of $1/ϵ$. This is computed with an expression similar to Eq. (19) of Ref. M. A. Bastarrachea-Magnani, S. Lerma-Hernández, J. G. Hirsch (2014), where we add an additional factor of $j$ to have units of $1/ϵ$ instead of $1/E$, and the integral is performed with a change of variable.

Arguments:

system should be generated with ClassicalDicke.ClassicalSystem.

j is the value of $j$

ε is the scaled energy $ϵ=E/j$

See Plotting the density of states for an example.

Dicke.ClassicalDicke.discriminant_of_q_solution — Method

function discriminant_of_q_solution(system::ClassicalSystems.ClassicalSystem; Q,P,p,ε)

Returns the discriminant of the second degree equation in $q$ given by

\[ h_\text{cl}(Q,q,P,p)=\epsilon,\]

where $h_\text{cl}$ is given by Eq. (5) of Ref. Saúl Pilatowsky-Cameo, David Villaseñor, Miguel A. Bastarrachea-Magnani, Sergio Lerma-Hernández, Lea F. Santos, Jorge G. Hirsch (2021).

Arguments:

system should be generated with ClassicalDicke.ClassicalSystem.

Q, P, p, and ε are the values of $Q$, $P$, $p$, and $\epsilon$, respectively.

Dicke.ClassicalDicke.energy_shell_volume — Method

function energy_shell_volume(system::ClassicalSystems.ClassicalSystem;ε)

Returns the volume of the classical energy shell in the phase space, that is, $\mathcal{V}(\mathcal{M}_ϵ)$ in Eq. (27) of Ref. Saúl Pilatowsky-Cameo, Jorge Chávez-Carlos, Miguel A. Bastarrachea-Magnani, Pavel Stránský, Sergio Lerma-Hernández, Lea F. Santos, Jorge G. Hirsch (2020).

Arguments:

system should be generated with ClassicalDicke.ClassicalSystem.

ε is the scaled energy $ϵ=E/j$

Dicke.ClassicalDicke.energy_width_of_coherent_state — Method

function energy_width_of_coherent_state(system::ClassicalSystems.ClassicalSystem,x,j::Real)

Returns the energy width $\sigma$ of the coherent state $\left | \mathbf{x}\right \rangle$, in units of $\epsilon$. This quantity is given by $\sigma_D/j$ with $\sigma_D$ as in App. A of Ref. Sergio Lerma-Hernández, Jorge Chávez-Carlos, Miguel A. Bastarrachea-Magnani, Lea F. Santos, Jorge G. Hirsch (2018).

Arguments:

system should be generated with ClassicalDicke.ClassicalSystem.

x is the coordinate $\mathbf{x}$ of the coherent state in the format [Q,q,P,p].

j is the value of $j$.

Dicke.ClassicalDicke.hamiltonian — Method

function hamiltonian(system::ClassicalSystems.ClassicalSystem)

Returns a classical Hamiltonian function h(x) where x=[Q,q,P,p], which is given by Eq. (5) of Ref. Saúl Pilatowsky-Cameo, David Villaseñor, Miguel A. Bastarrachea-Magnani, Sergio Lerma-Hernández, Lea F. Santos, Jorge G. Hirsch (2021).

Arguments:

system should be generated with ClassicalDicke.ClassicalSystem.

Dicke.ClassicalDicke.maximum_P_for_ε — Method

function maximum_P_for_ε(system::ClassicalSystems.ClassicalSystem,ε)

Computes the maximum value of the parameter $P$ accessible to the system at energy $\epsilon$.

Arguments:

system should be generated with ClassicalDicke.ClassicalSystem.

ε is the scaled energy $ϵ=E/j$.

Dicke.ClassicalDicke.maximum_Q_for_ε — Method

function maximum_Q_for_ε(system::ClassicalSystems.ClassicalSystem,ε)

See ClassicalDicke.maximum_P_for_ε.

Dicke.ClassicalDicke.minimum_energy — Method

function minimum_energy(system::ClassicalSystems.ClassicalSystem)

Returns the energy of the ground-state coordinate given by ClassicalDicke.minimum_energy_point.

Arguments:

system should be generated with ClassicalDicke.ClassicalSystem.

Note: This function currently only works for the supperadiant phase.

Dicke.ClassicalDicke.minimum_energy_point — Function

function minimum_energy_point(system::ClassicalSystems.ClassicalSystem,Qsign::Union{typeof(-),typeof(+)}=+)

Returns the ground-state coordinate, that is, $\mathbf{x}_\text{GS}$ below Eq. (7) of Ref. Saúl Pilatowsky-Cameo, David Villaseñor, Miguel A. Bastarrachea-Magnani, Sergio Lerma-Hernández, Lea F. Santos, Jorge G. Hirsch (2021).

Arguments:

system should be generated with ClassicalDicke.ClassicalSystem.

Qsign toggles the sign of the $Q$ coordinate, that is, + for $\mathbf{x}_\text{GS}$ and - for $\widetilde{\mathbf{x}}_\text{GS}$.

Note: This function currently only works for the superradiant phase.

Dicke.ClassicalDicke.minimum_ε_for — Method

function minimum_ε_for(system::ClassicalSystems.ClassicalSystem;Q=:nothing,q=:nothing,P=:nothing,p=:nothing)

Returns the minimum energy $\epsilon$ when constraining the system to three fixed values of the coordinates $Q$, $q$, $P$, $p$.

Arguments:

system should be generated with ClassicalDicke.ClassicalSystem.

You may pass either $(Q,q,P)$ or $(q,P,p)$. The other combinanations are not implemented.

This function can be especially useful to draw contours of the available phase space (see Drawing contours of the available phase space)

Dicke.ClassicalDicke.normal_frequency — Function

function normal_frequency(system::ClassicalSystems.ClassicalSystem,signo::Union{typeof(-),typeof(+)}=+)

Returns the ground-state normal frequency, that is, $\Omega_{\epsilon_{\text{GS}}}^{A,B}$ at the bottom of page 3 of Ref. Saúl Pilatowsky-Cameo, David Villaseñor, Miguel A. Bastarrachea-Magnani, Sergio Lerma-Hernández, Lea F. Santos, Jorge G. Hirsch (2021).

Arguments:

system should be generated with ClassicalDicke.ClassicalSystem.

signo is - for $\Omega^A$ and + for $\Omega^B$.

Note: This function currently only works for the supperadiant phase.

Dicke.ClassicalDicke.phase_space_dist_squared — Method

function phase_space_dist_squared(x,y)

Returns the phase-space distance $d_{\mathcal{M}}(\mathbf{x},\mathbf{y})$ (See App. C of Ref Saúl Pilatowsky-Cameo, David Villaseñor, Miguel A. Bastarrachea-Magnani, Sergio Lerma-Hernández, Lea F. Santos, Jorge G. Hirsch (2021)), where x and y are in the form [Q,q,P,p].

Dicke.ClassicalDicke.q_of_ε — Method

function q_of_ε(system::ClassicalSystems.ClassicalSystem;Q,P,p,ε,signo::Union{typeof(-),typeof(+)}=+,returnNaNonError=true)

Returns the solutions $q_\pm$ of the second degree equation in $q$ given by

\[ h_\text{cl}(Q,q,P,p)=\epsilon,\]

where $h_\text{cl}$ is given by Eq. (5) of Ref. Saúl Pilatowsky-Cameo, David Villaseñor, Miguel A. Bastarrachea-Magnani, Sergio Lerma-Hernández, Lea F. Santos, Jorge G. Hirsch (2021).

Arguments:

system should be generated with ClassicalDicke.ClassicalSystem.

Q, P, p, and ε are values of $Q$, $P$, $p$, and $\epsilon$, respectively.

signo is + for $q_+$ and - for $q_-$

If returnNaNonError is true, then NaN is returned if there are no solutions. If it is false, and error is raised.

Dicke.ClassicalDicke.q_sign — Function

function q_sign(system::ClassicalSystems.ClassicalSystem,x,ε=hamiltonian(system)(x))

Returns the sign of the root of the second degree equation in $q$ given by

\[ h_\text{cl}(Q,q,P,p)=\epsilon.\]

That is, this function returns + if q=x[2] ≈ q_of_ε(system;Q=x[1],P=x[3],p=x[4],ε=ε,signo=+) and returns - if q=x[2] ≈ q_of_ε(system;Q=x[1],P=x[3],p=x[4],ε=ε,signo=-).

Arguments:

system should be generated with ClassicalDicke.ClassicalSystem.

x is a vector in the form [Q,q,P,p].

ε should be the energy of x. If this is not passed, it is computed using hamiltonian(system).