DickeModel.ClassicalLMG

For examples of usage, go to Examples for ClassicalLMG.

DickeModel.ClassicalLMG.ClassicalLMGSystem — Type

struct ClassicalLMGSystem <: ClassicalSystems.ClassicalSystem

Subtype of ClassicalSystems.ClassicalSystem which represents the classical LMG model with the given parameters $Ω$ and $ξ$. See Eq. (2) of Ref. [12]. To generate this struct, use the constructor

    ClassicalLMGSystem(;Ω::Real,ξ::Real)

For example, system = ClassicalLMGSystem(Ω=1, ξ=1).

This struct may be passed to all functions in this module that require an instance of ClassicalLMG.ClassicalLMGSystem, as well as functions in other modules that require the abstract ClassicalSystems.ClassicalSystem, such as ClassicalSystems.integrate.

DickeModel.ClassicalLMG.hamiltonian — Function

function hamiltonian(system::ClassicalLMGSystem)

Returns a classical Hamiltonian function h(x) where x=[Q,P], which is given by Eq. (2) of Ref. [12].

Arguments

system should be generated with ClassicalLMG.ClassicalLMGSystem.

Phase space

DickeModel.ClassicalLMG.Point — Function

function Point(;Q::Real,P::Real)

Returns the list [Q,P]

function Point(system::ClassicalLMGSystem;
    Q::Real,
    ϵ::Real,
    sgn::Union{typeof(-),typeof(+)} = +)

Returns a list [Q,P], where P is calculated with P_of_ϵ. If there are no solutions for $P$, an error is raised.

DickeModel.ClassicalLMG.Pointθϕ — Function

function Pointθϕ(;θ::Real,ϕ::Real)

Returns a list [Q,P], where Q and P are calculated from θ and ϕ using PhaseSpaces.Q_of_θϕ and PhaseSpaces.P_of_θϕ.

Roots in P

DickeModel.ClassicalLMG.discriminant_of_P_solution — Function

function discriminant_of_P_solution(system::ClassicalLMGSystem,Q::Real,ϵ::Real)

Returns the discriminant of the second degree equation in $P$ given by

\[ h_\text{cl}(Q,P)=\epsilon,\]

where $h_\text{cl}$ is given by Eq. (2) of Ref. [12].

Arguments

system should be generated with ClassicalLMGSystem.

Q and ϵ are the values of $Q$ and $\epsilon$, respectively.

DickeModel.ClassicalLMG.P_of_ϵ — Function

function P_of_ϵ(system::ClassicalLMGSystem;
    Q::Real,
    ϵ::Real,
    sgn::Union{typeof(-),typeof(+)}=+,
    returnNaNonError::Bool=true)

Returns the solutions $P_\pm$ of the second degree equation in $P$ given by

\[ h_\text{cl}(Q,P)=\epsilon,\]

where $h_\text{cl}$ is given by Eq. (2) of Ref. [12].

Arguments

system should be generated with ClassicalLMGSystem.

Keyword arguments

Q and ϵ are values of $Q$ and $\epsilon$, respectively.

sgn is + for $P_+$ and - for $P_-$

If returnNaNonError is true, then NaN is returned if there are no solutions. If it is false, and error is raised.

DickeModel.ClassicalLMG.Point — Method

function Point(system::ClassicalLMGSystem;
    Q::Real,
    ϵ::Real,
    sgn::Union{typeof(-),typeof(+)} = +)

Returns a list [Q,P], where P is calculated with P_of_ϵ. If there are no solutions for $P$, an error is raised.

Energy minimizing

DickeModel.ClassicalLMG.minimum_ϵ_for — Function

function minimum_ϵ_for(system::ClassicalLMGSystem;
    Q::Union{Real,Nothing}=nothing,
    P::Union{Real,Nothing}=nothing)

Returns the minimum energy $\epsilon$ when constraining the system to one fixed value of the coordinates $Q$ or $P$.

Arguments

system should be generated with ClassicalLMGSystem.

Keyword arguments

You may pass either $P$ or $Q$.