Dicke.ClassicalSystems

Dicke.ClassicalSystems.integrate — Method

function integrate(system::ClassicalSystem;t::Real,
                   u₀::AbstractArray{<:Real, 1},
                   t₀=0.0::Real,
                   tol=1e-12::Real,
                   get_fundamental_matrix::Bool=false,
                   integrator_alg=TsitPap8(),
                   use_big_numbers::Bool=false,
                   integate_backwards::Bool=false,
                   kargs...)

This function integrates initial condition u₀ from t₀ to t under the Hamiltonian system determined by system, returning an instance of ODESolution from the package DifferentialEquations (See https://diffeq.sciml.ai/dev/basics/solution/).

Arguments:

system is an instance of ClassicalSystems.ClassicalSystem.

u₀ is an array which codifies the initial condition [Q,q,P,p] for ClassicalDicke and [Q,P] for ClassicalLMG.

t is the start time of the integration.

t₀ is the start of the integration (defaults to t₀ = 0.0) ClassicalSystems.ClassicalSystem.

tol is the tolerance for the integration, which determines both abstol and reltol in https://diffeq.sciml.ai/stable/basics/commonsolveropts/

get_fundamental_matrix determines whether to also compute the fundametal matrix of the system. If true, the result at each time is an ArrayPartition(x,Φ), so that x=result.x[2] retrieves the coordinate and Ψ=result.x[2] retrieves the fundamental matrix. Default is false. Note that the integration is consideribly slowed down if this parameter is set to true.

integrator_alg is the integration algorithm to use. Defaults to TsitPap8 (Tsitouras-Papakostas 8/7 Runge-Kutta method). See https://diffeq.sciml.ai/stable/solvers/ode_solve/#Full-List-of-Methods for other options.

use_big_numbers forces the integration to be performed with BigFloat instead of Float, allowing for infinite numerical precision, but hindering speed substantially. Defaults to false.

integate_backwards tells the integrator to integrate back in time, from -t₀ to -t. Defaults to false.

Additional kargs are passed to OrdinaryDiffEq.solve (see https://diffeq.sciml.ai/stable/basics/commonsolveropts/).

Dicke.ClassicalSystems.lyapunov_exponent — Method

function lyapunov_exponent(system::ClassicalSystem;kargs...)

Returns the maximal Lyapunov exponent for system.

Arguments

system is an instance of ClassicalSystems.ClassicalSystem.

kargs... are redirected to ClassicalSystems.integrate. In particular, you should pass u₀ and t.

PENDING EXAMPLE

Dicke.ClassicalSystems.lyapunov_spectrum — Method

function lyapunov_spectrum(system::ClassicalSystem;kargs...)

Same as ClassicalSystems.lyapunov_exponent, but returns the whole Lyapunov spectrum.

Dicke.ClassicalSystems.ClassicalSystem — Type

struct ClassicalSystem

This object represents a classical system that may be passed to multiple functions in this module. To generate a system, use ClassicalDicke.ClassicalSystem or use ClassicalDicke.ClassicalSystem.