Dicke.DickeBCE

Dicke.DickeBCE.H_BCE — Method

function H_BCE(system::QuantumSystem)

Returns a sparse matrix corresponding to the Dicke Hamiltonian (see any of the references. For example, Eq. (1) of [6]) in the efficient coherent basis.

Arguments:

system should be an instance of DickeBCE.QuantumSystem.

Dicke.DickeBCE.Jx — Method

function Jx(system::QuantumSystem)

Returns a sparse matrix representing the operator $\hat{J}_x$ in the efficient coherent basis.

Arguments:

system should be an instance of DickeBCE.QuantumSystem.

Dicke.DickeBCE.Jz — Method

function Jz(system::QuantumSystem)

Returns a sparse matrix representing the operator $\hat{J}_z$ in the efficient coherent basis.

Arguments:

system should be an instance of DickeBCE.QuantumSystem.

Dicke.DickeBCE.diagonalization — Method

function diagonalization(systemQ::QuantumSystem;
        load_cache = true,
        save_cache = true,
        cache_folder = joinpath(homedir(),"dicke_diagonalizations"),
        maxϵ::Real = 5.0,
        onlyload::Union{AbstractVector{<:Integer},Nothing} = nothing,
        only_eigenenergies = false,
        converged_tolerance=1e-3)

If load_cache = false or there are no saved diagonalizations in cache_folder, diagonalizes the Dicke Hamiltonian up to a maximum energy maxϵ. These eigenstates are guaranteed to be converged, with a tolerance determined by converged_tolerance. Numerical degeneracies are also corrected, to ensure that the eigenstates have parity +1 or -1. If only_eigenenergies is false (default), a tuple (eigenenergies,eigenstates) is returned, where eigenenergies is real vector containing the eigenenergies and eigenstates is a real matrix that contains the eigenstates as columns.

Arguments:

system should be an instance of DickeBCE.QuantumSystem.

load_cache is a boolean indicating whether to try to load from cache folder. Defaults to true.

save_cache determines if the results of a computed diagonalization are saved to the cache folder. Defaults to true.

cache_folder is the cache folder where diagonalizations are saved. Default is %HOME%/dicke_diagonalizations

maxϵ is the maximum energy up to which we diagonalize. Keep this number higher than the maximum converged regime you want.

onlyload may be a vector of integers, indicating the indices of the eigenstates to load, or the default value, nothing, indicates that all eigenstates should be loaded.

only_eigenenergies should be true if you only to load the eigenenergies. Defaults to false (return eigenstates and eigenenergies).

converged_tolerance determines how strict we are in saying that an eigenstate is converged. The default value 1e-3 is usually the best.

Dicke.DickeBCE.eigenenergies — Method

function eigenenergies(systemQ::QuantumSystem;args...)

Returns a vector containing the eigenenergies of the system. These are loaded or computed as in diagonalization.

Arguments:

system should be an instance of DickeBCE.QuantumSystem.

args... same as in diagonalization).

Dicke.DickeBCE.eigenstate_parities — Method

function eigenstate_parities(system::QuantumSystem,eigenstates)

Returns a vector of -1s and 1s contaning the parities of all of the eigenstates.

Arguments:

system should be an instance of DickeBCE.QuantumSystem.

eigenstates is the matrix of eigenstates. (See diagonalization).

Dicke.DickeBCE.factor_R_of_coherent_state — Method

function factor_R_of_coherent_state(system::QuantumSystem,
    x;
    eigenstates,
    eigenenergies,
    state = coherentBCE(system,x))

Computes $R$ for a coherent state, as defined in Eq. (30) of [9].

Arguments:

system should be an instance of DickeBCE.QuantumSystem.

x is an array [Q,q,P,p] representing the center of the coherent state.

eigenstates should be a matrix containing the eigenstates as columns.

eigenenergies is a list containing the eigenenergies.

state is a complex vector representing the coherent state $\left | \mathbf{x} \right \rangle$. If it is not passed, it is computed using DickeBCE.coherentBCE

Dicke.DickeBCE.participation_ratio — Method

function participation_ratio(state;eigenstates,eigenenergies=nothing,count_degeneracies=eigenenergies!=nothing,degentol=1e-5)

Returns the Participation Ratio (PR) of state in the eigenbasis. If count_degeneracies is true, then Eq. (7) of [9] is used. If count_degeneracies is false, then the inverse of Eq. (19) of [9] is used.

Arguments:

state should be a complex vector representing the state in the efficient coherent basis.

eigenstates should be a matrix containing the eigenstates.

eigenenergies should be passed if count_degeneracies is true. It is a list containing the eigenenergies.

count_degeneracies – whether to modify the PR definition to account for degeneracies (see above). Default is false if eigenenergies is not passed, else it is true.

degentol minimum energy separation below which two eigenstates are considered degenerate. Default is 1e-5.

Dicke.DickeBCE.survival_probability — Method

function survival_probability(t::Union{Real,AbstractArray{<:Real}};
    state,
    eigenstates,
    eigenenergies,
    normwarning=0.99)

PENDING

Arguments:

PENDING

Dicke.DickeBCE.Π_BCE — Method

function Π_BCE(system::QuantumSystem)

Returns a sparse matrix corresponding to the parity operator $\hat{\Pi}=e^{i\pi(\hat{a}^\dagger\hat{a}+ \hat{J}_z + j)}$ in the efficient coherent basis.

Arguments:

system should be an instance of DickeBCE.QuantumSystem.

Dicke.DickeBCE.QuantumSystem — Type

mutable struct QuantumSystem

This object represents the quantum Dicke model. It stores the parameters of the system, and it may be passed to multiple functions in this module. To generate a QuantumSystem, use QuantumSystem.

Dicke.DickeBCE.QuantumSystem — Method

function QuantumSystem(system::ClassicalDicke.ClassicalDickeSystem;j,Nmax=nothing)

Returns a QuantumSystem given $j$, $N_\text{max}$, $ω_0$, $ω$, and $γ$.

Arguments:

system should be generated with ClassicalDicke.ClassicalDickeSystem.

j is the value of $j$.

Nmax is the maximum excitation of the modified bosonic sector in the efficient coherent basis (see [2]). It may be omitted if there is a saved diagonalization. In that case, a call to diagonalization will populate this value with the greatest available in the saved cache.