Dicke.DickeBCE

Dicke.DickeBCE.QuantumDickeSystem — Type

mutable struct QuantumDickeSystem

This object represents the quantum Dicke model. It stores the parameters of the system, and it may be passed to multiple functions in this module. To generate it, use

function QuantumDickeSystem(classical_system::ClassicalDicke.ClassicalDickeSystem;
    j::Real,
    Nmax::Union{Integer,Nothing}=nothing)

Arguments:

classical_system should be generated with ClassicalDicke.ClassicalDickeSystem.

j is the value of $j$. It must be a positive half-integer.

Nmax - 1 is the maximum excitation of the modified bosonic sector in the efficient coherent basis (see [3], [4]). Nmax must be a positive integer. It may be omitted if there is a saved diagonalization. In that case, a call to diagonalization will populate this value with the greatest available in the saved cache.

Dicke.DickeBCE.QuantumDickeSystem — Method

function QuantumDickeSystem(;ω₀,ω,γ,j,Nmax=nothing)

Shorthand for QuantumDickeSystem(ClassicalDicke.ClassicalDickeSystem(ω₀=ω₀, ω=ω, γ=γ), j=j, Nmax=Nmax).

Dicke.DickeBCE.H_BCE — Method

function H_BCE(system::QuantumDickeSystem)

Returns a sparse matrix corresponding to the Dicke Hamiltonian (See Eq. (1) of [11]) in the efficient coherent basis. (See Ref. [3], [4])

Arguments:

system should be an instance of DickeBCE.QuantumDickeSystem.

Dicke.DickeBCE.Husimi — Method

function Husimi(args...;kargs...)

Computes abs2.(coherentOverlap(args...;kargs...)). The arguments and behavior are the same as coherentOverlap.

Dicke.DickeBCE.HusimiOfCoherent — Method

function HusimiOfCoherent(systemQ::QuantumDickeSystem,
    x::AbstractArray{<:Real,1},
    y::AbstractArray{<:Real,1})

Returns $\left | \left \langle \mathbf{x}\middle | \mathbf{y}\right \rangle \right | ^2$. Equation (3.14b) of Ref. [1] is used for the overlap of the Bloch coherent states.

Arguments:

systemQ should be an instance of DickeBCE.QuantumDickeSystem.

x and y are vectors in the form [Q, q, P, p].

Dicke.DickeBCE.Jx — Method

function Jx(system::QuantumDickeSystem)

Returns a sparse matrix representing the operator $\hat{J}_x$ in the efficient coherent basis.

Arguments:

system should be an instance of DickeBCE.QuantumDickeSystem.

Dicke.DickeBCE.Jz — Method

function Jz(system::QuantumDickeSystem)

Returns a sparse matrix representing the operator $\hat{J}_z$ in the efficient coherent basis.

Arguments:

system should be an instance of DickeBCE.QuantumDickeSystem.

Dicke.DickeBCE.Wigner — Method

function Wigner(systemQ::QuantumDickeSystem,
    state::AbstractArray{<:Number},
    points::Vector{<:AbstractVector{<:Real}})

Evaluates the Wigner function of state (which is a vector in the BCE) in all points = [[Q1,q1,P1,p1], [Q2,q2,P2,p2], ...], returning an Array with the results.

Note: This function has not been thoroughly tested. It is based on Ref. [7].

Dicke.DickeBCE.WignerProjQP — Method

function WignerProjQP(systemQ::QuantumDickeSystem,
    state::AbstractArray{<:Number},
    ptsqp::Vector{<:AbstractVector{<:Real}})

Evaluates Wigner function, integrated over the atomic variables $Q,P$, of state in all ptsqp = [[q1,p1], [q2,p2], ...], returning an array with the results.

Note: This function has not been thoroughly tested. It is based on Ref. [7].

Dicke.DickeBCE.WignerProjqp — Method

function WignerProjqp(systemQ::QuantumDickeSystem,
    states::AbstractArray{<:AbstractArray{<:Number}},
    ptsQP::Vector{<:AbstractVector{<:Real}};
    show_progress::Bool=true)

Evaluates Wigner function, integrated over the bosonic variables $q,p$, of each of the states in states (which is a vector of complex vectors) in all ptsQP = [[Q1,P1], [Q2,P2], ...], returning an array with the results. See this example. If show_progress = true it shows a progress bar.

Note: This function has not been thoroughly tested. It is based on Ref. [7].

Dicke.DickeBCE.coherentBCE! — Method

function coherentBCE!(system::QuantumDickeSystem,
            x::AbstractArray{<:Real,1},
            data::AbstractVector{<:Number};
            normwarning=0.99,
            add_state_to_data::Bool=false,
            extra_phase::Complex=1.0+0im,
            tol::Real=1e-6)

This function computes the coefficients of a coherent state centered at x in the BCE, and stores the result in data. (See Ref. [14])

Arguments:

system should be an instance of DickeBCE.QuantumDickeSystem.

x is a vector with the coordinates [Q, q, P, p].

data should be a complex vector of length dimension(systemQ).

normwarning is a tolerance (defaults to 0.99). If the norm of the resulting coherent state is below this number, a warning will be thrown. This usually happens if Nmax is too small for the energy regime you are working on.

add_state_to_data is a boolean. If it is true, the coefficients will be added to data. If it is false (default), this function will override any preexisting values in `data. This is useful if you want to add many coherents states together without having to allocate that much memory.

add_state_to_data is a boolean. If true, the coefficients will be added to data, and if false (default), this function will override any preexisting values in `data.

extra_phase is a complex number that multiplies the resulting state overall. Defaults to 1.

tol is a numerical tolerance. If tol=0, then all the coefficients all computed. However, if ε = 1 - tol > 0.0, then a some coefficients at the tail of the distribution and whose squared norm sum does not exceed ε, will be treated as 0.0. This allows to significally reduce computation time, but introduces a numerical error of order ε. The default is 1e-6. See this example and Ref. [8].

Dicke.DickeBCE.coherentBCE — Method

function coherentBCE(system::QuantumDickeSystem,
    x::AbstractVector{<:Real};
    normwarning::Real=0.99,
    extra_phase::Complex=1.0+0im,
    tol::Real=0.0)

Calls coherentBCE! passing data as a new vector and returns the result.

Dicke.DickeBCE.coherentOverlap — Method

function coherentOverlap(systemQ::QuantumDickeSystem,
    x::AbstractVector{<:Real},
    states::AbstractMatrix{<:Number};
    tol::Real=1e-6,
    normwarning::Real=0.99,
    datacache::Union{AbstractVector{<:Complex},Nothing}=nothing)

Same as coherentOverlap(..., state,...) below, but instead of one state, it allows states to be a matrix with multiple states as columns. It is functionally equivalent to

[coherentOverlap(..., states[:,k], ...) for k in 1:size(states)[2]]

but it is (much) faster. The result is stored in datacache if provided.

Dicke.DickeBCE.coherentOverlap — Method

function coherentOverlap(systemQ::QuantumDickeSystem,
    x::AbstractVector{<:Real},
    state::AbstractVector{<:Number};
    tol::Real=1e-6,
    normwarning::Real=0.99,
    datacache::Union{AbstractArray{Complex{Float64},1},Nothing}=nothing)

Returns the overlap $\left \langle \mathbf{x}\middle | \text{state} \right \rangle$, where $\left | \mathbf{x} \right \rangle$ is a coherent state centered at x.

Arguments:

systemQ should be an instance of DickeBCE.QuantumDickeSystem.

x is a vector in the form [Q, q, P, p].

state is a vector representing $\left | \text{state} \right \rangle$ in the BCE.

tol is a numerical tolerance. If tol=0, then all the products of coefficients are computed. However, if ε = 1 - tol > 0.0, then a some coefficients of the coherent state at the tail of the distribution and whose squared norm sum does not exceed ε, will be treated as 0.0. This allows to significally reduce computation time, but introduces a numerical error of order ε. The default is 1e-6. See this example and Ref. [8].

normwarning is a tolerance (defaults to 0.99). If the norm of the coherent state is below this number, a warning will be thrown. This usually happens if Nmax is too small for the energy regime you are working on.

datacache in an array where to store the result, or nothing for the result to be returned.

Dicke.DickeBCE.diagonalization — Method

function diagonalization(systemQ::QuantumDickeSystem;
        load_cache = true,
        save_cache = true,
        cache_folder = joinpath(homedir(),"dicke_diagonalizations"),
        maxϵ::Real = 5.0,
        onlyload::Union{AbstractVector{<:Integer},Nothing} = nothing,
        only_eigenenergies = false,
        verbose::Bool = true,
        converged_tolerance=1e-3)

If load_cache = false or there are no saved diagonalizations in cache_folder, diagonalizes the Dicke Hamiltonian up to a maximum energy maxϵ. The resulting eigenstates are guaranteed to be converged, with a tolerance determined by converged_tolerance. Numerical degeneracies are also corrected, to ensure that the eigenstates have parity +1 or -1. If only_eigenenergies is false (default), a tuple (eigenenergies,eigenstates) is returned, where eigenenergies is real vector containing the eigenenergies and eigenstates is a real matrix that contains the eigenstates as columns.

Arguments:

system should be an instance of DickeBCE.QuantumDickeSystem.

load_cache is a boolean indicating whether to try to load from cache folder. Defaults to true.

save_cache determines if the results of a computed diagonalization are saved to the cache folder. Defaults to true.

cache_folder is the cache folder where diagonalizations are saved. Default is %HOME%/dicke_diagonalizations

maxϵ is the maximum energy up to which we diagonalize. Keep this number higher than the maximum converged regime you want.

onlyload may be a vector of integers, indicating the indices of the eigenstates to load, or the default value, nothing, indicates that all eigenstates should be loaded.

only_eigenenergies should be true if you only to load the eigenenergies. Defaults to false (return eigenstates and eigenenergies).

verbose is true (default) if you want to see info messages.

converged_tolerance determines how strict we are in saying that an eigenstate is converged (see [4]). The default value 1e-3 is usually the best.

Dicke.DickeBCE.dimension — Method

function dimension(system::QuantumDickeSystem)

Returns the integer $(2j + 1)\times N_\text{max}$, which gives the dimension of the Hilbert space.

Dicke.DickeBCE.eigenenergies — Method

function eigenenergies(systemQ::QuantumDickeSystem;args...)

Returns a vector containing the eigenenergies of the system. These are loaded or computed as in diagonalization.

Arguments:

system should be an instance of DickeBCE.QuantumDickeSystem.

args... – same as in diagonalization.

Dicke.DickeBCE.eigenstate_parities — Method

function eigenstate_parities(system::QuantumDickeSystem,eigenstates)

Returns a vector of -1s and 1s contaning the parities of all of the eigenstates.

Arguments:

system should be an instance of DickeBCE.QuantumDickeSystem.

eigenstates is the matrix of eigenstates. (See diagonalization).

Dicke.DickeBCE.evolve — Method

function evolve(t::Union{Real,AbstractArray{<:Real}};
    state::AbstractVector{<:Number},
    eigenstates::AbstractMatrix{<:Number},
    eigenenergies::AbstractVector{<:Real},
    normwarning::Real=0.99)

Computes the evolution $\left | \text{state}(t)\right \rangle$ of state under the Dicke Hamiltonian. The result is returned in the BCE. at the times given.

Arguments:

t may be a single time, or a vector of times. In the first case, a vector will be returned, and in the second, a matrix with the state corresponding to each time as a column.

state is a vector representing the state in the BCE.

eigenstates and eigenenergies are the ones returned by diagonalization.

normwarning is a tolerance (defaults to 0.99). If the norm of the state in the eigenbasis is below this number, a warning will be thrown. This usually happens if Nmax is too small for the energy regime you are working on.

Dicke.DickeBCE.factor_R_of_coherent_state — Method

function factor_R_of_coherent_state(system::QuantumDickeSystem,
        x::AbstractVector{<:Real};
        eigenstates::AbstractMatrix{<:Number},
        eigenenergies::AbstractVector{<:Real},
        state::AbstractVector{<:Number} = coherentBCE(system,x))

Computes $R$ for a coherent state, as defined in Eq. (30) of [14].

Arguments:

system should be an instance of DickeBCE.QuantumDickeSystem.

x is an array [Q,q,P,p] representing the center of the coherent state.

eigenstates should be a matrix containing the eigenstates as columns.

eigenenergies is a list containing the eigenenergies.

state is a complex vector representing the coherent state $\left | \mathbf{x} \right \rangle$. If it is not passed, it is computed using DickeBCE.coherentBCE

Dicke.DickeBCE.participation_ratio — Method

function participation_ratio(state::AbstractVector{<:Number};
    eigenstates::AbstractMatrix{<:Number},
    eigenenergies::Union{AbstractVector{<:Real},Nothing}=nothing,
    count_degeneracies::Bool=eigenenergies!=nothing,
    degentol::Real=1e-5)

Returns the Participation Ratio (PR) of state in the eigenbasis. If count_degeneracies is true, then Eq. (7) of [14] is used. If count_degeneracies is false, then the inverse of Eq. (19) of [14] is used.

Arguments:

state should be a complex vector representing the state in the efficient coherent basis.

eigenstates should be a matrix containing the eigenstates.

eigenenergies should be passed if count_degeneracies is true. It is a list containing the eigenenergies.

count_degeneracies – whether to modify the PR definition to account for degeneracies (see above). Default is false if eigenenergies is not passed, else it is true.

degentol minimum energy separation below which two eigenstates are considered degenerate. Default is 1e-5.

Dicke.DickeBCE.randomCoherentStatesInEnergyShell — Method

function randomCoherentStatesInEnergyShell(
    systemQ::QuantumDickeSystem;
    ϵ::Real,
    N::Integer,
    dt::Real=3,
    tol::Real=1e-6,
    cachedata::Union{AbstractVector{<:Complex},Nothing}=nothing)

Samples N points from the energy shell at ϵ using ClassicalDicke.classicalPathRandomSampler, and then constructs the N-cat state of all the coherent states centered at those points.

Dicke.DickeBCE.randomState! — Method

function randomState!(systemQ::QuantumDickeSystem,
    data::AbstractVecOrMat{<:Number};
    kargs...)

Generates as many random states as columns in data saving the result to data. The general form of the random states in the eigenbasis is $\left | R \right \rangle = \sum_k c_k \left | E_k \right \rangle$, where the $c_k$ are computed using random_cₖ_generator. You may select only positive or negative parity eigenstates (see below).

Arguments

systemQ should be an instance of DickeBCE.QuantumDickeSystem.

data should be a vector or matrix of numbers, where the result is stored. It must have as many columns as random numbers you want and dimension(systemQ) rows.

Required keyword arguments

Please pass all the required keyword arguments of random_cₖ_generator

eigenstates and eigenenergies shoud be passed, being the ones returned by diagonalization.

Optional keyword arguments

All the optional keyword arguments of random_cₖ_generator may be passed.

tol may be a real number which determines the tolerance for the state convergence. A value tol=0 builds the state with all of the eigenstates, and a positive number cuts the tails of the energy envelope up to that number. Higher values are faster, but decrease precision (Defaults to 1e-6)

parity may be +, -, or nothing (default). If it is not nothing, the random will be composed of only the eigenstates with this parity.

parities should be passed if parity is not nothing. It should be a vector of -1s and 1s containing the parities of all of the eigenstates. If it is not passed, it will be computed with eigenstate_parities, but this is slow. If you plan to use this function repeatedly, you should precompute parities by calling eigenstate_parities yourself and then pass it to avoid repeated calls.

Dicke.DickeBCE.randomState — Method

function randomState(systemQ::QuantumDickeSystem;
                    number_of_states=1,
                    kargs...)

Generates number_of_states random states, returning them as columns in a matrix (or a vector if number_of_states). kargs are redirected to randomState!: see that function for more details.

Dicke.DickeBCE.random_cₖ_generator — Method

function random_cₖ_generator(systemQ::QuantumDickeSystem;
                            kargs...)

Generates coefficients to build random states, which are given by

\[ c_k(\epsilon_k) \sim e^{i \theta_k} \sqrt{\frac{r_k\rho(\epsilon_k)}{\nu(\epsilon_k)}}.\]

(See Refs. [6], [14], and [11]). It returns a function cₖ(ϵ;number=1,cache=nothing), which produces n=number coefficients, storing them in cache (an n-vector), if given.

The parameters of the equation above are determined as follows:

Required keyword arguments

To determine $\rho(\epsilon)$ above, pass either:

σ and ϵ, both real, in which case $\rho$ will be taken as a normal distribution centered at ϵ with standard deviation σ,

OR envelope as any of Distributions.UnivariateDistribution.

To determine $r_k$, pass either:

ensemble = :GUE, which takes $r_k$ from an exponential distribution Exponential(0.91) or ensemble = :GOE, which takes $r_k$ from a $\chi^2$ distribution with one degree of freedom,

OR rₖ_distribution as any of Distributions.UnivariateDistribution.

To determine $\theta_k$, pass either:

phases = :real, which takes $\theta_k$ to be random from $\{ -\pi, \pi \}$ or phases = :complex, which takes $\theta_k$ to be uniform in $[0, 2\pi)$,

OR ensemble = :GOE, which automatically sets phases = :real, or ensemble = :GUE, which automatically sets phases = :complex.

Optional keyword arguments

divide_by_DoS is a boolean. If true (default), leaves the density of states, $\nu(\epsilon_k)$ in the equation above. A value of false removes it.

Dicke.DickeBCE.survival_probability — Method

function survival_probability(t::Union{Real,AbstractArray{<:Real}};
    state::AbstractVector{<:Number},
    eigenstates::AbstractMatrix{<:Number},
    eigenenergies::AbstractVector{<:Real},
    normwarning::Real=0.99)

Computes the survival probability $\left | \left \langle \text{state}\middle | \text{state}(t)\right \rangle \right | ^2$ at the times given.

Arguments:

t may be a single time, or a vector of times. In the first case, a number will be returned, and in the second, a vector.

state is a vector representing the state in the BCE.

eigenstates and eigenenergies are the ones returned by diagonalization.

normwarning is a tolerance (defaults to 0.99). If the norm of the state in the eigenbasis is below this number, a warning will be thrown. This usually happens if Nmax is too small for the energy regime you are working on.

Dicke.DickeBCE.Π_BCE — Method

function Π_BCE(system::QuantumDickeSystem)

Returns a sparse matrix corresponding to the parity operator $\hat{\Pi}=e^{i\pi(\hat{a}^\dagger\hat{a}+ \hat{J}_z + j)}$ in the efficient coherent basis.

Arguments:

system should be an instance of DickeBCE.QuantumDickeSystem.